# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
...
      "AspectRatio": {
        "description": "Frame shape. Stored as an enum rather than a free string because the\nrender pipeline (PHASE 13) resolves it to exact pixel dimensions, and\n\"9 : 16\" versus \"9:16\" would be two products' worth of bugs.",
        "enum": ["9:16", "16:9", "1:1", "4:5"],
        "title": "AspectRatio",
        "type": "string"
      },
...
      "ErrorResponse": {
        "description": "Uniform error envelope returned by every failing endpoint.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": ["error"],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ProjectPurpose": {
        "description": "What the video is for (§10.9, §16).\n\nFeeds the creative prompt, so these are not cosmetic: a launch film and a\nmarketplace listing want different structures from the same product.",
        "enum": [
          "LAUNCH",
          "ECOMMERCE_LISTING",
          "SOCIAL_AD",
          "BRAND_STORY",
          "FEATURE_HIGHLIGHT",
          "TUTORIAL",
          "OTHER"
        ],
        "title": "ProjectPurpose",
        "type": "string"
      },
      "ProjectResponse": {
        "properties": {
          "aspect_ratio": {
            "$ref": "#/components/schemas/AspectRatio"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "integer"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product_id": {
            "format": "uuid",
            "title": "Product Id",
            "type": "string"
          },
          "purpose": {
            "$ref": "#/components/schemas/ProjectPurpose"
          },
          "quality_mode": {
            "$ref": "#/components/schemas/QualityMode"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "style": {
            "$ref": "#/components/schemas/VideoStyle"
          },
          "target_audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Audience"
          },
          "target_platform": {
            "$ref": "#/components/schemas/TargetPlatform"
          }
        },
        "required": [
          "id",
          "product_id",
          "name",
          "purpose",
          "target_platform",
          "target_audience",
          "language",
          "aspect_ratio",
          "duration_seconds",
          "style",
          "quality_mode",
          "status",
          "failure_reason",
          "created_at"
        ],
        "title": "ProjectResponse",
        "type": "object"
      },
      "ProjectStatus": {
        "description": "Where a video project has got to (§10.9, §105).\n\nLonger than the product's machine because a project *is* the pipeline: each\nstate names a stage that can fail on its own and be resumed on its own.",
        "enum": [
          "DRAFT",
          "ANALYZING",
          "CREATIVE_PLANNING",
          "SCRIPTING",
          "STORYBOARDING",
          "GENERATING",
          "COMPOSITING",
          "QC",
          "READY",
          "FAILED",
          "ARCHIVED"
        ],
        "title": "ProjectStatus",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API for AI Product Video Studio.\n\nLong-running AI work (analysis, shot generation, TTS, render, QC) is **never**\nperformed inside a request. Those endpoints validate, create a `GenerationJob`,\nenqueue it and return `202 Accepted` with a `job_id` to poll (taskbook §22).\n\nEvery failure returns the same envelope:\n\n```json\n{\"error\": {\"code\": \"PROJECT_NOT_FOUND\", \"message\": \"...\", \"request_id\": \"...\"}}\n```\n\n`code` comes from a closed set (§65) and is the field clients should branch on.",
    "title": "AI Product Video Studio API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/projects": {
      "get": {
        "operationId": "list_projects_api_v1_workspaces__workspace_id__projects_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "product_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "in": "query",
            "name": "project_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProjectStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "title": "Response List Projects Api V1 Workspaces  Workspace Id  Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List projects",
        "tags": ["projects"]
      },
      "post": {
        "description": "Create a project against a product. Always starts in DRAFT (§105).",
        "operationId": "create_project_api_v1_workspaces__workspace_id__projects_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Start a project",
        "tags": ["projects"]
      }
    },
    "/api/v1/workspaces/{workspace_id}/projects/{project_id}": {
      "get": {
        "operationId": "get_project_api_v1_workspaces__workspace_id__projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get a project",
        "tags": ["projects"]
      },
      "patch": {
        "description": "Change the brief, while the project is still ahead of generation.\n\nRefused afterwards: changing the duration once shots exist would silently\ninvalidate work already paid for. Move the project back a stage first.",
        "operationId": "update_project_api_v1_workspaces__workspace_id__projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Edit the brief",
        "tags": ["projects"]
      }
    },
    "/api/v1/workspaces/{workspace_id}/projects/{project_id}/archive": {
      "post": {
        "operationId": "archive_project_api_v1_workspaces__workspace_id__projects__project_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Archive a project",
        "tags": ["projects"]
      }
    },
...
  }
}